style: incapsulated AdminApi
