+++ admin-bot/src/main/kotlin/com/github/heheteam/adminbot/states/CourseInfoState.kt
@@ -1,7 +1,7 @@
 package com.github.heheteam.adminbot.states
 
+import com.github.heheteam.adminbot.AdminApi
 import com.github.heheteam.adminbot.CourseStatistics
-import com.github.heheteam.adminbot.CourseStatisticsComposer
 import com.github.heheteam.adminbot.Keyboards
 import com.github.heheteam.adminbot.formatters.CourseStatisticsFormatter
 import com.github.heheteam.commonlib.Course
@@ -15,21 +15,17 @@ import dev.inmo.tgbotapi.types.chat.User
 import kotlinx.coroutines.flow.first
 
 class CourseInfoState(override val context: User, val course: Course) :
-  BotState<Unit, CourseStatistics, CourseStatisticsComposer> {
-  override suspend fun readUserInput(bot: BehaviourContext, service: CourseStatisticsComposer) =
-    Unit
+  BotState<Unit, CourseStatistics, AdminApi> {
+  override suspend fun readUserInput(bot: BehaviourContext, service: AdminApi) = Unit
 
-  override fun computeNewState(
-    service: CourseStatisticsComposer,
-    input: Unit,
-  ): Pair<State, CourseStatistics> {
+  override fun computeNewState(service: AdminApi, input: Unit): Pair<State, CourseStatistics> {
     val stats = service.getCourseStatistics(course.id)
     return Pair(MenuState(context), stats)
   }
 
   override suspend fun sendResponse(
     bot: BehaviourContext,
-    service: CourseStatisticsComposer,
+    service: AdminApi,
     response: CourseStatistics,
   ) {
     val statsMessage =
