+++ admin-bot/src/main/kotlin/com/github/heheteam/adminbot/states/RemoveTeacherState.kt
@@ -1,6 +1,6 @@
 package com.github.heheteam.adminbot.states
 
-import com.github.heheteam.adminbot.AdminCore
+import com.github.heheteam.adminbot.AdminApi
 import com.github.heheteam.adminbot.Dialogues.manyIdsAlreadyDoNotExistForTeacherRemoving
 import com.github.heheteam.adminbot.Dialogues.manyIdsAreGoodForTeacherRemoving
 import com.github.heheteam.adminbot.Dialogues.manyTeacherIdsDoNotExist
@@ -25,17 +25,17 @@ import dev.inmo.tgbotapi.types.chat.User
 import dev.inmo.tgbotapi.types.message.abstracts.AccessibleMessage
 
 class RemoveTeacherState(override val context: User, val course: Course, val courseName: String) :
-  BotStateWithHandlers<String, List<String>, AdminCore> {
+  BotStateWithHandlers<String, List<String>, AdminApi> {
 
   val sentMessages = mutableListOf<AccessibleMessage>()
 
-  override suspend fun outro(bot: BehaviourContext, service: AdminCore) {
+  override suspend fun outro(bot: BehaviourContext, service: AdminApi) {
     // No special cleanup needed
   }
 
   override suspend fun intro(
     bot: BehaviourContext,
-    service: AdminCore,
+    service: AdminApi,
     updateHandlersController: UpdateHandlersController<() -> Unit, String, Any>,
   ) {
     val introMessage =
@@ -49,7 +49,7 @@ class RemoveTeacherState(override val context: User, val course: Course, val cou
   }
 
   @Suppress("LongMethod", "CyclomaticComplexMethod") // wild legacy, fix later
-  override fun computeNewState(service: AdminCore, input: String): Pair<State, List<String>> {
+  override fun computeNewState(service: AdminApi, input: String): Pair<State, List<String>> {
     if (input == "/stop") {
       return Pair(MenuState(context), emptyList())
     }
@@ -113,7 +113,7 @@ class RemoveTeacherState(override val context: User, val course: Course, val cou
 
   override suspend fun sendResponse(
     bot: BehaviourContext,
-    service: AdminCore,
+    service: AdminApi,
     response: List<String>,
   ) {
     sentMessages.forEach { bot.delete(it) }
